Move ServerEndpoint{,s} to play.core.server
