+++ framework/src/play-integration-test/src/test/scala/play/it/http/FlashCookieSpec.scala
@@ -14,6 +14,7 @@ import play.api.mvc.Results._
 import play.api.mvc._
 import play.api.routing.Router
 import play.api.test._
+import play.core.server.ServerEndpoint
 import play.it.test.{ EndpointIntegrationSpecification, OkHttpEndpointSupport }
 
 import scala.collection.JavaConverters
