+++ framework/src/play-integration-test/src/test/scala/play/it/test/EndpointIntegrationSpecification.scala
@@ -8,7 +8,8 @@ import org.specs2.execute.{ AsResult, PendingUntilFixed, Result, ResultExecution
 import org.specs2.mutable.SpecLike
 import org.specs2.specification.core.Fragment
 
-import play.api.test.{ ApplicationFactory, ApplicationFactories, ServerEndpoint, ServerEndpointRecipe }
+import play.api.test.{ ApplicationFactories, ApplicationFactory, ServerEndpointRecipe }
+import play.core.server.ServerEndpoint
 
 /**
  * Mixin class for integration tests that want to run over different
