+++ framework/src/play-microbenchmark/src/test/scala/play/microbenchmark/it/HelloWorldBenchmark.scala
@@ -9,7 +9,8 @@ import java.util.concurrent.TimeUnit
 import okhttp3.{ OkHttpClient, Protocol, Request, Response }
 import org.openjdk.jmh.annotations._
 import play.api.mvc.Results
-import play.api.test.{ ApplicationFactory, ServerEndpoint, ServerEndpointRecipe }
+import play.api.test.{ ApplicationFactory, ServerEndpointRecipe }
+import play.core.server.ServerEndpoint
 import play.microbenchmark.it.HelloWorldBenchmark.ThreadState
 
 import scala.util.Random
