+++ framework/src/play-server/src/main/scala/play/core/server/ServerEndpoints.scala
@@ -2,7 +2,7 @@
  * Copyright (C) 2009-2018 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package play.api.test
+package play.core.server
 
 import akka.annotation.ApiMayChange
